refactor(forks/lstar): merge store module into containers
Move AttestationSignatureEntry and Store next to the other lstar
container types, then delete the standalone store module so the fork
exposes a single container namespace.
