+++ packages/testing/src/consensus_testing/test_types/block_spec.py
@@ -22,9 +22,9 @@ from lean_spec.spec.forks.lstar.containers import (
     SignedBlock,
     SingleMessageAggregate,
     State,
+    Store,
 )
 from lean_spec.spec.forks.lstar.spec import LstarSpec
-from lean_spec.spec.forks.lstar.store import Store
 from lean_spec.spec.ssz import ByteList512KiB, Bytes32
 
 from ..keys import XmssKeyManager, create_dummy_signature
