+++ packages/testing/src/consensus_testing/test_types/gossip_attestation_spec.py
@@ -5,9 +5,8 @@ from __future__ import annotations
 from lean_spec.base import CamelModel
 from lean_spec.spec.crypto.merkleization import hash_tree_root
 from lean_spec.spec.forks import Checkpoint, Slot, ValidatorIndex
-from lean_spec.spec.forks.lstar.containers import AttestationData, Block, SignedAttestation
+from lean_spec.spec.forks.lstar.containers import AttestationData, Block, SignedAttestation, Store
 from lean_spec.spec.forks.lstar.spec import LstarSpec
-from lean_spec.spec.forks.lstar.store import Store
 from lean_spec.spec.ssz import Bytes32
 
 from ..keys import XmssKeyManager, create_dummy_signature
