+++ src/lean_spec/spec/forks/lstar/__init__.py
@@ -1,7 +1,6 @@
 """Lstar fork"""
 
-from .containers import State
+from .containers import AttestationSignatureEntry, State
 from .spec import LstarStore as Store
-from .store import AttestationSignatureEntry
 
 __all__ = ["AttestationSignatureEntry", "State", "Store"]
